update libarchive-static to 1.1.0
The previous update got fig installing on Ubuntu 22.04, but there was an
incompaitibility with the very old version of libarchive bundled with
the libarchive-static gem. This new version has the latest libarchive
and zlib.

This update breaks MacOS support as the libarchive-static 1.1.0 gem does
not compile on MacOS at the moment. Windows support is not yet tested. A
future update will provide support for both of those platforms.