Simplify workflow eliminating half-baked detect-native-gems.
The native gem detection scheme worked some of the time but not all
the time, but it remained because it was the only one that ran the
tests and ran on multiple os's.  That's moved into `package`.

`detect` also had the illusion of supporting multiple
platform/architectures but it really didn't. Probably better to
support that in the test-gem jobs anyway.