Teach file protocol how to deal with the linked split repo.
Tests failed b/c Find.find() doesn't follow symlinks on traversal and
doesn't offer the option. A real implementation isn't hard but it's
not as easy as just swapping out the real path everywhere (need to
avoid graph cycles). This takes advantage of insider knowledge of the
environment in which we need it to work checks to see if it's given a
symlink for the repo path, and then traverses the realpath instead.