Fix failing test.
This worked previously because of a side-effect in the "file" protocol
that creates the target directory during an upload using the
equivalent of `mkdir -p`.  This has the effect of also creating the
target remote dir.

Since the target remote is a symlink now in order to simulate the
linked artifactory repo, the side-effect creation doesn't do the right
thing yielding test failures.