Remove tebako from standard build workflow.
While tebako would be great to use, it looks like its development has
stalled and something changed around the beginning of Sep 2025 where
macos14 builds fail in a boost step building folly.

https://github.com/tamatebako/tebako/actions/runs/17370390435

It's a great tool and idea, but has a lot of moving parts that need
tending that it's apparently not getting.