Simplify steps required to publish a new release.
Releasing needs to be simple, because people aren't going to be firing
off new releases of fig on the daily.  I personally found myself
forgetting steps and having to rebase/amend, move tags, etc.

This change tightly couples the tag to the version file change,
automating tag creation from the version file committed to
master.  The decision for when to release, which semver component to
bump and what value to use, etc., all remains very manual.  A better
future choice is conventional commits and release-please.